M    configure M    include/ag_debug.h M    include/ag_button.h M    configure.in M    src/ag_debug.cc M    src/ag_table.cc M    src/ag_glsurface.cc M    src/ag_messageobject.cc M    src/ag_button.cc M    src/Makefile.am M    autom4te.cache/traces.0 M    autom4te.cache/output.0

+++ antargisgui/trunk/src/ag_table.cc
@@ -124,7 +124,7 @@ void AGTable::arrange()
 	    AGWidget *wd=children[mx+my*w];
 	    if(wd)
 	      {
-		wd->setWidth((width()-mfx)*cols[mx]/xw);
+		wd->setWidth((int)((width()-mfx)*cols[mx]/xw));
 	      }
 	  }
 	if(rows[my]>0.0)
@@ -132,7 +132,7 @@ void AGTable::arrange()
 	    AGWidget *wd=children[mx+my*w];
 	    if(wd)
 	      {
-		wd->setHeight((height()-mfy)*rows[my]/yw);
+		wd->setHeight((int)((height()-mfy)*rows[my]/yw));
 	      }
 	  }
       }
